requirements.txt
flask
pillow
opencv-python
numpy
easyocr
gunicorn